MB-37420: Match wildcards in subjectAltNames
https://github.com/erlang/otp/pull/1667

After internal discussion it was decided we need to enable wildcards
in SAN for Mad-Hatter hence the patch for erlang 20.

Erlang team added an option customize_hostname_check to enable SAN
wildcards in erlang 21 but we decided not to backport their
solution because of the following reasons:
- the patch would become more complex and more risky;
- appling soultion from otp team would require extra work to be done
  in order to enable SAN wildcards for distribution (because they
  only provide an option, so distribution needs to start using this
  option);
- we will have to fix other then distribution clients that we use as
  well (mostly http clients I guess).

Related OTP commits:
https://github.com/erlang/otp/commit/f821c91cebe0cee22c1c6e0a9dfe45d4e9b5f129
https://github.com/erlang/otp/commit/66ed8bc404d56cf7277ac34226a8d605740f492e

When we migrate to erlang 21 or higher we need to reevaluate the
option of using of the otp team fix. I suspect it might be still
easier to just reapply this patch though.